Map working with markers.  DB updated to accept floats

+++ db/migrate/20151009172301_create_benches.rb
@@ -2,8 +2,8 @@ class CreateBenches < ActiveRecord::Migration
   def change
     create_table :benches do |t|
       t.text :description, null: false
-      t.integer :lat, null: false
-      t.integer :lng, null: false
+      t.float :lat, null: false
+      t.float :lng, null: false
 
       t.timestamps null: false
     end
